Added the (dropdown) Wishlist section
After clicking on "Add to Wishlist," the item will be added to the Wishlist section. In the Wishlist, the user can either add the item to the cart or remove it. The user can view their Wishlist and add items to the cart with just one click, without needing to navigate to another page.