Added the (dropdown) Wishlist - updated
Fixes Issue #53 - Sending a new pull request for this, as I have improved its CSS compared to that in the previous pull request. (After clicking on "Add to Wishlist," the item will be added to the Wishlist section. In the Wishlist, the user can either add the item to the cart or remove it. The user can view their Wishlist and add items to the cart, without needing to navigate to another page.)